+++ include/ruckig/utils.hpp
@@ -41,7 +41,7 @@ inline std::string join(const Vector& array, bool high_precision = false) {
 
 
 //! Integrate with constant jerk for duration t. Returns new position, new velocity, and new acceleration.
-inline std::tuple<double, double, double> integrate(double t, double p0, double v0, double a0, double j) {
+inline std::tuple<float, float, float> integrate(float t, float p0, float v0, float a0, float j) {
     return std::make_tuple(
         p0 + t * (v0 + t * (a0 / 2 + t * j / 6)),
         v0 + t * (a0 + t * j / 2),
